+++ libfreerdp/core/info.c
@@ -314,7 +314,7 @@ BOOL rdp_read_info_packet(wStream* s, rdpSettings* settings)
 	Stream_Read_UINT16(s, cbAlternateShell); /* cbAlternateShell */
 	Stream_Read_UINT16(s, cbWorkingDir); /* cbWorkingDir */
 
-	if (Stream_GetRemainingLength(s) < cbDomain + 2)
+	if (Stream_GetRemainingLength(s) < (size_t) (cbDomain + 2))
 		return FALSE;
 
 	if (cbDomain > 0)
@@ -324,7 +324,7 @@ BOOL rdp_read_info_packet(wStream* s, rdpSettings* settings)
 	}
 	Stream_Seek(s, 2);
 
-	if (Stream_GetRemainingLength(s) < cbUserName + 2)
+	if (Stream_GetRemainingLength(s) < (size_t) (cbUserName + 2))
 		return FALSE;
 
 	if (cbUserName > 0)
@@ -334,7 +334,7 @@ BOOL rdp_read_info_packet(wStream* s, rdpSettings* settings)
 	}
 	Stream_Seek(s, 2);
 
-	if (Stream_GetRemainingLength(s) < cbPassword + 2)
+	if (Stream_GetRemainingLength(s) < (size_t) (cbPassword + 2))
 		return FALSE;
 
 	if (cbPassword > 0)
@@ -344,7 +344,7 @@ BOOL rdp_read_info_packet(wStream* s, rdpSettings* settings)
 	}
 	Stream_Seek(s, 2);
 
-	if (Stream_GetRemainingLength(s) < cbAlternateShell + 2)
+	if (Stream_GetRemainingLength(s) < (size_t) (cbAlternateShell + 2))
 		return FALSE;
 
 	if (cbAlternateShell > 0)
@@ -354,7 +354,7 @@ BOOL rdp_read_info_packet(wStream* s, rdpSettings* settings)
 	}
 	Stream_Seek(s, 2);
 
-	if (Stream_GetRemainingLength(s) < cbWorkingDir + 2)
+	if (Stream_GetRemainingLength(s) < (size_t) (cbWorkingDir + 2))
 		return FALSE;
 
 	if (cbWorkingDir > 0)
